requirements.txt
django
pyjwt
